# EVM gas for memory tiny optimization