--- a/ethcore/src/evm/evm.rs
+++ b/ethcore/src/evm/evm.rs
@@ -107,6 +107,11 @@ pub trait CostType: ops::Mul<Output=Self> + ops::Div<Output=Self> + ops::Add<Out
 	fn overflow_add(self, other: Self) -> (Self, bool);
 	/// Multiple with overflow
 	fn overflow_mul(self, other: Self) -> (Self, bool);
+	/// Divide with overflow
+	fn overflow_div(self, other: Self) -> (Self, bool);
+	/// Single-step full multiplication and division: `self*other/div`
+	/// Should not overflow on intermediate steps
+	fn overflow_mul_div(self, other: Self, div: Self) -> (Self, bool);
 }
 
 impl CostType for U256 {
@@ -128,6 +133,21 @@ impl CostType for U256 {
 
 	fn overflow_mul(self, other: Self) -> (Self, bool) {
 		Uint::overflowing_mul(self, other)
+	}
+
+	fn overflow_div(self, other: Self) -> (Self, bool) {
+		Uint::overflowing_div(self, other)
+	}
+
+	fn overflow_mul_div(self, other: Self, div: Self) -> (Self, bool) {
+		let x = self.full_mul(other);
+		let (U512(parts), o) = Uint::overflowing_div(x, U512::from(div));
+		let overflow = (parts[4] | parts[5] | parts[6] | parts[7]) > 0;
+
+		(
+			U256([parts[0], parts[1], parts[2], parts[3]]),
+			o | overflow
+		)
 	}
 }
 
@@ -154,6 +174,18 @@ impl CostType for usize {
 	fn overflow_mul(self, other: Self) -> (Self, bool) {
 		self.overflowing_mul(other)
 	}
+
+	fn overflow_div(self, other: Self) -> (Self, bool) {
+		self.overflowing_div(other)
+	}
+
+	fn overflow_mul_div(self, other: Self, div: Self) -> (Self, bool) {
+		let (c, o) = U128::from(self).overflowing_mul(U128::from(other));
+		let (U128(parts), o1) = c.overflowing_div(U128::from(div));
+		let result = parts[0] as usize;
+		let overflow = o | o1 | (parts[1] > 0) | (parts[0] > result as u64);
+		(result, overflow)
+	}
 }
 
 /// Evm interface
@@ -164,3 +196,41 @@ pub trait Evm {
 	/// to compute the final gas left.
 	fn exec(&mut self, params: ActionParams, ext: &mut Ext) -> Result<GasLeft>;
 }
+
+
+#[test]
+fn should_calculate_overflow_mul_div_without_overflow() {
+	// given
+	let num = 10_000_000;
+
+	// when
+	let (res1, o1) = U256::from(num).overflow_mul_div(U256::from(num), U256::from(num));
+	let (res2, o2) = num.overflow_mul_div(num, num);
+
+	// then
+	assert_eq!(res1, U256::from(num));
+	assert!(!o1);
+	assert_eq!(res2, num);
+	assert!(!o2);
+}
+
+#[test]
+fn should_calculate_overflow_mul_div_with_overflow() {
+	// given
+	let max = ::std::u64::MAX;
+	let num1 = U256([max, max, max, max]);
+	let num2 = ::std::usize::MAX;
+
+	// when
+	let (res1, o1) = num1.overflow_mul_div(num1, num1 - U256::from(2));
+	let (res2, o2) = num2.overflow_mul_div(num2, num2 - 2);
+
+	// then
+	// (x+2)^2/x = (x^2 + 4x + 4)/x = x + 4 + 4/x ~ (MAX-2) + 4 + 0 = 1
+	assert_eq!(res2, 1);
+	assert!(o2);
+
+	assert_eq!(res1, U256::from(1));
+	assert!(o1);
+}
+