Benches:
Comparison with `ethvm`:

Example 1 (loop)
ethvm: 1.50573 seconds.
usize: Time: 0.452586528s
U256: Time: 0.631101462s

Example 2 (rng)
ethvm: 7.7506 seconds.
usize: Time: 1.999621423s
U256: Time: 2.413986383s